chore: Update Playwright configuration for Chromium
This commit updates the Playwright configuration for the Chromium project. It adds additional properties to the `use` object, including `baseURL` set to `http://localhost:4002` and `channel` set to `chrome`. It also adds a `port` property to the `webServer` object, set to `4002`. These changes are made to improve the configuration for running tests with Playwright in the Chromium browser.